+++ lib/Ast.mli
@@ -193,8 +193,17 @@ val is_simple : Conf.t -> (expression xt -> int) -> expression xt -> bool
     applications of other simple expressions. *)
 
 val exposed_left_typ : core_type -> bool
+(** [exposed_left_typ typ] holds iff the given type begins with a [<]
+    character. *)
+
+(** [exposed_right_*] predicates hold iff the supplied argument ends with a
+    [>] character. *)
 
 val exposed_right_typ : core_type -> bool
+
+val exposed_right_label_declaration : label_declaration -> bool
+
+val exposed_right_row_field : row_field -> bool
 
 (** 'Classes' of expressions which are parenthesized differently. *)
 type cls = Let_match | Match | Non_apply | Sequence | Then | ThenElse